research/steve/requirements.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
cffi==1.15.0
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.24
dataclasses==0.8
decorator==5.1.0
dm-tree==0.1.6
fasteners==0.16.3
future==0.18.2
gast==0.2.2
glfw==2.4.0
google-pasta==0.2.0
grpcio==1.41.1
gym==0.21.0
h5py==3.1.0
imageio==2.10.3
importlib-metadata==4.8.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mujoco-py==2.0.2.13
numpy==1.18.5
nvidia-cublas==11.2.1.74
nvidia-cuda-cupti==11.1.69
nvidia-cuda-nvcc==11.1.74
nvidia-cuda-nvrtc==11.1.105
nvidia-cuda-runtime==11.1.74
nvidia-cudnn==8.0.4.30
nvidia-cufft==10.3.0.74
nvidia-curand==10.2.2.74
nvidia-cusolver==11.0.0.74
nvidia-cusparse==11.2.0.275
nvidia-dali-cuda110==0.27.0
nvidia-dali-nvtf-plugin==0.27.0+nv20.11
nvidia-horovod==0.20.2+nv20.11
nvidia-nccl==2.8.2
nvidia-pyindex==1.0.9
nvidia-tensorboard==1.15.0+nv20.11
nvidia-tensorflow==1.15.4+nv20.11
nvidia-tensorrt==7.2.1.6
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
pkg_resources==0.0.0
portalocker==2.3.2
protobuf==3.19.1
psutil==5.8.0
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.5.4
six==1.16.0
tensorboard @ https://pypi.ngc.nvidia.com/tensorboard/tensorboard-1.15.0-py2.py3-none-any.whl
tensorflow-estimator==1.15.1
tensorflow-probability==0.7.0
termcolor==1.1.0
typing-extensions==3.10.0.2
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
